Fix warning

+++ bettery.c
@@ -8,8 +8,8 @@
 // https://docs.google.com/spreadsheets/d/1qrFdMTo0ZrqtlGUoafeB3hplhU3GzDnVWuUK4M9OgNo/edit?usp=sharing
 #define MULTI (float)0.443
 
-#define VOLTAGE_MIN (double)2.0
-#define VOLTAGE_MAX (double)3.3
+#define VOLTAGE_MIN 2.0
+#define VOLTAGE_MAX 3.3
 
 uint8 getBatteryVoltageZCL(uint16 millivolts) {
     uint8 volt8 = (uint8)(millivolts / 100);
@@ -26,4 +26,4 @@ uint16 getBatteryVoltage(void) {
     return (uint16)(rawADC * MULTI);
 }
 
-uint8 getBatteryRemainingPercentageZCL(uint16 millivolts) { return mapRange(VOLTAGE_MIN, VOLTAGE_MAX, 0, 200, millivolts); }
+uint8 getBatteryRemainingPercentageZCL(uint16 millivolts) { return (uint8) mapRange(VOLTAGE_MIN, VOLTAGE_MAX, 0.0, 200.0, millivolts); }
